refactor: extract ai/ml utilities
