+++ src/core/managers/extended/ai_ml/ai_manager.py
@@ -10,10 +10,10 @@ from typing import Any, Dict, List, Optional
 import logging
 
 from src.core.base_manager import BaseManager
-from .configuration import load_ai_config
-from .constants import DEFAULT_AI_MANAGER_CONFIG
-from .lifecycle import AgentLifecycle
-from .metrics import MetricsCollector
+from src.extended.ai_ml.config import load_ai_config
+from src.extended.ai_ml.constants import DEFAULT_AI_MANAGER_CONFIG
+from src.extended.ai_ml.lifecycle import AgentLifecycle
+from src.extended.ai_ml.metrics import MetricsCollector
 
 logger = logging.getLogger(__name__)
