+++ src/extended/__init__.py
@@ -1,0 +1,1 @@
+"""Extended utilities package."""
