+++ src/extended/ai_ml/__init__.py
@@ -1,0 +1,17 @@
+"""Extended AI/ML shared utilities."""
+
+from .config import AIConfig, load_ai_config
+from .constants import DEFAULT_AI_MANAGER_CONFIG
+from .lifecycle import AgentLifecycle
+from .metrics import MetricsCollector
+from .orchestrator import OrchestrationTask, SystemHealth
+
+__all__ = [
+    "AIConfig",
+    "load_ai_config",
+    "DEFAULT_AI_MANAGER_CONFIG",
+    "AgentLifecycle",
+    "MetricsCollector",
+    "OrchestrationTask",
+    "SystemHealth",
+]
